Fix unit tests using Obj::set<size_t>

+++ test/test_query.cpp
@@ -9922,10 +9922,10 @@ TEST(Query_IntOrQueryPerformance)
     auto ints_col_key = table->add_column(type_Int, "ints");
     auto nullable_ints_col_key = table->add_column(type_Int, "nullable_ints", true);
 
-    const size_t null_frequency = 1000;
-    size_t num_nulls_added = 0;
-    size_t limit = 100000;
-    for (size_t i = 0; i < limit; ++i) {
+    const int null_frequency = 1000;
+    int num_nulls_added = 0;
+    int limit = 100000;
+    for (int i = 0; i < limit; ++i) {
         if (i % null_frequency == 0) {
             auto o = table->create_object().set_all(i);
             o.set_null(nullable_ints_col_key);
@@ -9936,14 +9936,14 @@ TEST(Query_IntOrQueryPerformance)
         }
     }
 
-    auto run_queries = [&](size_t num_matches) {
+    auto run_queries = [&](int num_matches) {
         // std::cout << "num_matches: " << num_matches << std::endl;
         Query q_ints = table->column<Int>(ints_col_key) == -1;
         Query q_nullables =
             (table->column<Int>(nullable_ints_col_key) == -1).Or().equal(nullable_ints_col_key, realm::null());
-        for (size_t i = 0; i < num_matches; ++i) {
-            q_ints = q_ints.Or().equal(ints_col_key, int64_t(i));
-            q_nullables = q_nullables.Or().equal(nullable_ints_col_key, int64_t(i));
+        for (int i = 0; i < num_matches; ++i) {
+            q_ints = q_ints.Or().equal(ints_col_key, i);
+            q_nullables = q_nullables.Or().equal(nullable_ints_col_key, i);
         }
 
         auto before = std::chrono::steady_clock().now();
@@ -9995,15 +9995,15 @@ TEST(Query_IntFindInNextLeaf)
     auto col_id = table->add_column(type_Int, "id");
 
     // num_misses > MAX_BPNODE_SIZE to check results on other leafs
-    constexpr size_t num_misses = 1000 * 2 + 10;
-    for (size_t i = 0; i < num_misses; i++) {
+    constexpr int num_misses = 1000 * 2 + 10;
+    for (int i = 0; i < num_misses; i++) {
         table->create_object().set(col_id, i % 10);
     }
     table->create_object().set(col_id, 20);
 
     auto check_results = [&]() {
-        for (size_t i = 0; i < 10; ++i) {
-            Query qi = table->where().equal(col_id, int64_t(i));
+        for (int i = 0; i < 10; ++i) {
+            Query qi = table->where().equal(col_id, i);
             CHECK_EQUAL(qi.count(), num_misses / 10);
         }
         Query q20 = table->where().equal(col_id, 20);
